serialization working correctly for all graph representations

+++ src/operables/mathematics/ArithmeticOperation.h
@@ -112,7 +112,7 @@ inline void ArithmeticOperation::serialize(Archive& ar,
 }
 
 // Associate some type with a version number
-CEREAL_CLASS_VERSION( ArithmeticOperation, 1 );
+CEREAL_CLASS_VERSION( ArithmeticOperation, (int)1 );
 
 // Include any archives you plan on using with your type before you register it
 // Note that this could be done in any other location so long as it was prior
